fix thneed, rename cl to _cl
